Move some more things around.
